Capability manager format
This commit format the information of capability manager. It fixes the warning messages and includes the file in table of contect of authorization enabler.
It changes the name of files also. All file/directory names should be lower case. Use - for imporving reading of the file names. e.g. capability-manager.
Fix image tag in .md file.

Signed-off-by: b-rohit <rohitbohra9@gmail.com>